Revert "Internal: Make tabs active state controllable [ED-20986]" (#33454)
Reverts elementor/elementor#33431
<!--start_gitstream_placeholder-->
### ✨ PR Description
Purpose: Revert controls for tabs active state and remove preview
handler features that were previously added.
Main changes:
- Removed tab preview handler script and related event listeners for
Navigator integration
- Restored original tabs handler implementation with inline utility
functions
- Fixed element attributes by removing x-ref bindings from tab
components

_Generated by LinearB AI and added by gitStream._
<sub>AI-generated content may contain inaccuracies. Please verify before
using.
💡 **Tip:** You can customize your AI Description using **Guidelines**
[Learn
how](https://docs.gitstream.cm/automation-actions/#describe-changes)</sub>
<!--end_gitstream_placeholder-->
