+++ modules/atomic-widgets/assets/js/editor/atomic-element-types/atomic-tab/create-atomic-tab-view.js
@@ -5,7 +5,7 @@ const createAtomicTabView = () => {
 		attributes() {
 			const attributes = super.attributes();
 
-			return { 'x-bind': 'tab', 'x-ref': this.model.id, ...attributes };
+			return { 'x-bind': 'tab', ...attributes };
 		}
 	};
 };
