+++ modules/atomic-widgets/elements/atomic-tabs/atomic-tab.php
@@ -178,7 +178,6 @@ class Atomic_Tab extends Atomic_Element_Base {
 			'tabindex' => $is_active ? '0' : '-1',
 			'aria-selected' => $is_active ? 'true' : 'false',
 			'x-bind' => 'tab',
-			'x-ref' => $this->get_id(),
 			'id' => Atomic_Tabs::get_tab_id( $tabs_id, $index ),
 			'aria-controls' => Atomic_Tabs::get_tab_content_id( $tabs_id, $index ),
 		];
