+++ modules/atomic-widgets/elements/atomic-tabs/atomic-tabs-handler.js
@@ -1,12 +1,16 @@
 import { register } from '@elementor/frontend-handlers';
 import { Alpine } from '@elementor/alpinejs';
-import { TAB_ELEMENT_TYPE, TAB_CONTENT_ELEMENT_TYPE, getTabId, getTabContentId, getIndex, getNextTab } from './utils';
+
+const TAB_ELEMENT_TYPE = 'e-tab';
+const TAB_CONTENT_ELEMENT_TYPE = 'e-tab-content';
 
 const SELECTED_CLASS = 'e--selected';
+const NAVIGATE_UP_KEYS = [ 'ArrowUp', 'ArrowLeft' ];
+const NAVIGATE_DOWN_KEYS = [ 'ArrowDown', 'ArrowRight' ];
 
 register( {
 	elementType: 'e-tabs',
-	id: 'e-tabs-handler',
+	uniqueId: 'e-tabs-handler',
 	callback: ( { element, settings } ) => {
 		Alpine.data( 'atomicTabs', () => ( {
 			activeTab: settings[ 'default-active-tab' ],
@@ -20,7 +24,7 @@ register( {
 				':id'() {
 					const index = getIndex( this.$el, TAB_ELEMENT_TYPE );
 
-					return getTabId( element.dataset.id, index );
+					return getTabId( index );
 				},
 				'@click'() {
 					const id = this.$el.id;
@@ -57,7 +61,7 @@ register( {
 				':aria-controls'() {
 					const index = getIndex( this.$el, TAB_ELEMENT_TYPE );
 
-					return getTabContentId( element.dataset.id, index );
+					return getTabContentId( index );
 				},
 			},
 
@@ -65,11 +69,11 @@ register( {
 				':aria-labelledby'() {
 					const index = getIndex( this.$el, TAB_CONTENT_ELEMENT_TYPE );
 
-					return getTabId( element.dataset.id, index );
+					return getTabId( index );
 				},
 				'x-show'() {
 					const index = getIndex( this.$el, TAB_CONTENT_ELEMENT_TYPE );
-					const tabId = getTabId( element.dataset.id, index );
+					const tabId = getTabId( index );
 
 					const isActive = this.activeTab === tabId;
 
@@ -82,10 +86,49 @@ register( {
 				':id'() {
 					const index = getIndex( this.$el, TAB_CONTENT_ELEMENT_TYPE );
 
-					return getTabContentId( element.dataset.id, index );
+					return getTabContentId( index );
 				},
 			},
 		} ) );
+
+		const getTabId = ( tabIndex ) => {
+			const tabsId = element.dataset.id;
+			return `${ tabsId }-tab-${ tabIndex }`;
+		};
+
+		const getTabContentId = ( tabIndex ) => {
+			const tabsId = element.dataset.id;
+			return `${ tabsId }-tab-content-${ tabIndex }`;
+		};
+
+		const getChildren = ( el, elementType ) => {
+			const parent = el.parentElement;
+
+			return Array.from( parent.children ).filter( ( child ) => {
+				return child.dataset.element_type === elementType;
+			} );
+		};
+
+		const getIndex = ( el, elementType ) => {
+			const children = getChildren( el, elementType );
+
+			return children.indexOf( el );
+		};
+
+		const getNextTab = ( key, tab ) => {
+			const tabs = getChildren( tab, TAB_ELEMENT_TYPE );
+			const tabsLength = tabs.length;
+
+			const currentIndex = getIndex( tab, TAB_ELEMENT_TYPE );
+
+			if ( NAVIGATE_DOWN_KEYS.includes( key ) ) {
+				return tabs[ ( currentIndex + 1 ) % tabsLength ];
+			}
+
+			if ( NAVIGATE_UP_KEYS.includes( key ) ) {
+				return tabs[ ( currentIndex - 1 + tabsLength ) % tabsLength ];
+			}
+		};
 	},
 } );
