+++ modules/atomic-widgets/elements/atomic-tabs/atomic-tabs.php
@@ -16,7 +16,6 @@ use Elementor\Modules\AtomicWidgets\Controls\Types\Elements\Tabs_Control;
 use Elementor\Core\Utils\Collection;
 use Elementor\Modules\AtomicWidgets\Loader\Frontend_Assets_Loader;
 use Elementor\Utils;
-use Elementor\Plugin;
 
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -154,10 +153,6 @@ class Atomic_Tabs extends Atomic_Element_Base {
 	}
 
 	public function get_script_depends() {
-		if ( Plugin::$instance->preview->is_preview_mode() ) {
-			return [ 'elementor-tabs-handler', 'elementor-tabs-preview-handler' ];
-		}
-
 		return [ 'elementor-tabs-handler' ];
 	}
 
@@ -168,14 +163,6 @@ class Atomic_Tabs extends Atomic_Element_Base {
 		wp_register_script(
 			'elementor-tabs-handler',
 			"{$assets_url}js/tabs-handler{$min_suffix}.js",
-			[ Frontend_Assets_Loader::FRONTEND_HANDLERS_HANDLE, Frontend_Assets_Loader::ALPINEJS_HANDLE ],
-			ELEMENTOR_VERSION,
-			true
-		);
-
-		wp_register_script(
-			'elementor-tabs-preview-handler',
-			"{$assets_url}js/tabs-preview-handler{$min_suffix}.js",
 			[ Frontend_Assets_Loader::FRONTEND_HANDLERS_HANDLE, Frontend_Assets_Loader::ALPINEJS_HANDLE ],
 			ELEMENTOR_VERSION,
 			true
